requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# LLM
openai>=1.0.0

# Notion
notion-client>=2.2.1

# 工具库
python-dotenv==1.0.0
pydantic>=2.0.0

# 测试
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
